+++ app/routes/changelog.ts
@@ -12,6 +12,6 @@ export default class ChangelogRoute extends Route {
   }
 
   @action didTransition() {
-    this.renderer.tryResetScrollPosition();
+    this.renderer.trySetScrollPosition();
   }
 }